requirements.txt
-e .
aws-cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_rds
aws_cdk.aws_efs
aws_cdk.aws_ssm
aws_cdk.custom_resources
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_certificatemanager
aws_cdk.aws_autoscaling
aws_cdk.aws_autoscaling_hooktargets
aws_cdk.aws_fsx